목적에 맞게 컨트롤러 분리 (레벨업, 골드획득)
