+++ Assets/_Dev/Scripts/Controller/LevelUpController.cs.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: de450fd5f54e0574b962401a62a49f1c
+guid: 85779886e9fabb445b91f54e5d2b60f0
 MonoImporter:
   externalObjects: {}
   serializedVersion: 2
